move seer types to separate module
